promote nori tools to main jar
